[1/2] Move InternVL-based processors
